requirements.txt
Pillow==10.3.0
pyperclip==1.8.2
tiktoken==0.7.0
ttkthemes==3.2.2